+++ internal/commands/fetch_test.go
@@ -50,10 +50,13 @@ func TestFetchCommand_FetchByTMIDOrName(t *testing.T) {
 	rm.On("Get", remotes.NewRemoteSpec("r1")).Return(r, nil)
 	setUpVersionsForFetchByTMIDOrName(r)
 
-	r.On("Fetch", "manufacturer/mpn/v1.0.0-20231205123243-c49617d2e4fc.tm.json").Return("manufacturer/mpn/v1.0.0-20231205123243-c49617d2e4fc.tm.json", []byte("{}"), nil)
-	r.On("Fetch", "manufacturer/mpn/folder/sub/v1.0.0-20231205123243-c49617d2e4fc.tm.json").Return("manufacturer/mpn/folder/sub/v1.0.0-20231205123243-c49617d2e4fc.tm.json", []byte("{}"), nil)
-	r.On("Fetch", "author/manufacturer/mpn/v1.0.0-20231205123243-c49617d2e4fc.tm.json").Return("author/manufacturer/mpn/v1.0.0-20231205123243-c49617d2e4fc.tm.json", []byte("{}"), nil)
-	r.On("Fetch", "author/manufacturer/mpn/folder/sub/v1.0.0-20231205123243-c49617d2e4fc.tm.json").Return("author/manufacturer/mpn/folder/sub/v1.0.0-20231205123243-c49617d2e4fc.tm.json", []byte("{}"), nil)
+	r.On("Fetch", "manufacturer/mpn/v1.0.0-20231205123243-c49617d2e4fc.tm.json").Return("manufacturer/mpn/v1.0.0-20231205123243-c49617d2e4fc.tm.json", []byte("{\"ver\":\"v1.0.0\"}"), nil)
+	r.On("Fetch", "manufacturer/mpn/folder/sub/v1.0.0-20231205123243-c49617d2e4fc.tm.json").Return("manufacturer/mpn/folder/sub/v1.0.0-20231205123243-c49617d2e4fc.tm.json", []byte("{\"ver\":\"v1.0.0\"}"), nil)
+	r.On("Fetch", "author/manufacturer/mpn/v1.0.0-20231205123243-c49617d2e4fc.tm.json").Return("author/manufacturer/mpn/v1.0.0-20231205123243-c49617d2e4fc.tm.json", []byte("{\"ver\":\"v1.0.0\"}"), nil)
+	r.On("Fetch", "author/manufacturer/mpn/v1.0.4-20231206123243-d49617d2e4fc.tm.json").Return("author/manufacturer/mpn/v1.0.4-20231206123243-d49617d2e4fc.tm.json", []byte("{\"ver\":\"v1.0.4\"}"), nil)
+	r.On("Fetch", "author/manufacturer/mpn/v1.2.3-20231207153243-e49617d2e4ff.tm.json").Return("author/manufacturer/mpn/v1.2.3-20231207153243-e49617d2e4ff.tm.json", []byte("{\"ver\":\"v1.2.3\"}"), nil)
+	r.On("Fetch", "author/manufacturer/mpn/v2.0.0-20231208123243-f49617d2e4fc.tm.json").Return("author/manufacturer/mpn/v2.0.0-20231208123243-f49617d2e4fc.tm.json", []byte("{\"ver\":\"v2.0.0\"}"), nil)
+	r.On("Fetch", "author/manufacturer/mpn/folder/sub/v1.0.0-20231205123243-c49617d2e4fc.tm.json").Return("author/manufacturer/mpn/folder/sub/v1.0.0-20231205123243-c49617d2e4fc.tm.json", []byte("{\"ver\":\"v1.0.0\"}"), nil)
 
 	f := NewFetchCommand(rm)
 
@@ -61,22 +64,28 @@ func TestFetchCommand_FetchByTMIDOrName(t *testing.T) {
 		in         string
 		expErr     bool
 		expErrText string
-		expBLen    int
+		expVer     string
 	}{
-		{"", true, "Invalid name format:  - Must be NAME[:SEMVER|DIGEST]", 0},
-		{"manufacturer", true, "Invalid name format: manufacturer - Must be NAME[:SEMVER|DIGEST]", 0},
-		{"manufacturer/mpn", false, "", 2},
-		{"manufacturer/mpn/v1.0.0-20231205123243-c49617d2e4fc.tm.json", false, "", 2},
-		{"manufacturer/mpn/folder/sub/v1.0.0-20231205123243-c49617d2e4fc.tm.json", false, "", 2},
-		{"author/manufacturer/mpn", false, "", 2},
-		{"author/manufacturer/mpn:v1.0.0", false, "", 2},
-		{"author/manufacturer/mpn:1.0.0", false, "", 2},
-		{"author/manufacturer/mpn:c49617d2e4fc", false, "", 2},
-		{"author/manufacturer/mpn/v1.0.0-20231205123243-c49617d2e4fc.tm.json", false, "", 2},
-		{"author/manufacturer/mpn/folder/sub", false, "", 2},
-		{"author/manufacturer/mpn/folder/sub:v1.0.0", false, "", 2},
-		{"author/manufacturer/mpn/folder/sub:c49617d2e4fc", false, "", 2},
-		{"author/manufacturer/mpn/folder/sub/v1.0.0-20231205123243-c49617d2e4fc.tm.json", false, "", 2},
+		{"", true, "Invalid name format:  - Must be NAME[:SEMVER|DIGEST]", ""},
+		{"manufacturer", true, "Invalid name format: manufacturer - Must be NAME[:SEMVER|DIGEST]", ""},
+		{"manufacturer/mpn", false, "", ""},
+		{"manufacturer/mpn/v1.0.0-20231205123243-c49617d2e4fc.tm.json", false, "", "v1.0.0"},
+		{"manufacturer/mpn/folder/sub/v1.0.0-20231205123243-c49617d2e4fc.tm.json", false, "", "v1.0.0"},
+		{"author/manufacturer/mpn", false, "", "v2.0.0"},
+		{"author/manufacturer/mpn:c49617d2e4fc", false, "", "v1.0.0"},
+		{"author/manufacturer/mpn/v1.0.0-20231205123243-c49617d2e4fc.tm.json", false, "", "v1.0.0"},
+		{"author/manufacturer/mpn:v1.0.0", false, "", "v1.0.0"},
+		{"author/manufacturer/mpn:1.0.0", false, "", "v1.0.0"},
+		{"author/manufacturer/mpn:v1.0", false, "", "v1.0.4"},
+		{"author/manufacturer/mpn:1.3", true, "No version 1.3 found", ""},
+		{"author/manufacturer/mpn:1.1", true, "No version 1.1 found", ""},
+		{"author/manufacturer/mpn:1.2", false, "", "v1.2.3"},
+		{"author/manufacturer/mpn:3", true, "No version 3 found", ""},
+		{"author/manufacturer/mpn:v1", false, "", "v1.2.3"},
+		{"author/manufacturer/mpn/folder/sub", false, "", "v1.0.0"},
+		{"author/manufacturer/mpn/folder/sub:v1.0.0", false, "", "v1.0.0"},
+		{"author/manufacturer/mpn/folder/sub:c49617d2e4fc", false, "", "v1.0.0"},
+		{"author/manufacturer/mpn/folder/sub/v1.0.0-20231205123243-c49617d2e4fc.tm.json", false, "", "v1.0.0"},
 	}
 
 	for _, test := range tests {
@@ -86,7 +95,7 @@ func TestFetchCommand_FetchByTMIDOrName(t *testing.T) {
 			assert.ErrorContains(t, err, test.expErrText, "Unexpected error in FetchByTMIDOrName(%s)", test.in)
 		} else {
 			assert.NoError(t, err, "Expected no error in FetchByTMIDOrName(%s)", test.in)
-			assert.Equal(t, test.expBLen, len(b), "Unexpected result length in FetchByTMIDOrName(%s)", test.in)
+			assert.True(t, bytes.Contains(b, []byte(test.expVer)), "FetchByTMIDOrName(%s) result does not contain %s. Got: %s", test.in, test.expVer, string(b))
 		}
 	}
 }
@@ -120,6 +129,60 @@ func setUpVersionsForFetchByTMIDOrName(r *remotes.MockRemote) {
 					Version:   model.Version{Model: "v1.0.0"},
 					TMID:      "author/manufacturer/mpn/v1.0.0-20231205123243-c49617d2e4fc.tm.json",
 					Digest:    "c49617d2e4fc",
+					TimeStamp: "20231205123243",
+				},
+				FoundIn: model.FoundSource{RemoteName: "r1"},
+			},
+			{
+				TOCVersion: model.TOCVersion{
+					Version:   model.Version{Model: "v1.0.4"},
+					TMID:      "author/manufacturer/mpn/v1.0.4-20231206123243-d49617d2e4fc.tm.json",
+					Digest:    "d49617d2e4fc",
+					TimeStamp: "20231206123243",
+				},
+				FoundIn: model.FoundSource{RemoteName: "r1"},
+			},
+			{
+				TOCVersion: model.TOCVersion{
+					Version:   model.Version{Model: "v1.2.0"},
+					TMID:      "author/manufacturer/mpn/v1.2.0-20231207163243-e49617d2e4fc.tm.json",
+					Digest:    "e49617d2e4fc",
+					TimeStamp: "20231207163243", // this is on purpose more recent by timestamp than the latest semver (v.1.2.3)
+				},
+				FoundIn: model.FoundSource{RemoteName: "r1"},
+			},
+			{
+				TOCVersion: model.TOCVersion{
+					Version:   model.Version{Model: "v1.2.1"},
+					TMID:      "author/manufacturer/mpn/v1.2.1-20231207133243-e49617d2e4fd.tm.json",
+					Digest:    "e49617d2e4fd",
+					TimeStamp: "20231207133243",
+				},
+				FoundIn: model.FoundSource{RemoteName: "r1"},
+			},
+			{
+				TOCVersion: model.TOCVersion{
+					Version:   model.Version{Model: "v1.2.2"},
+					TMID:      "author/manufacturer/mpn/v1.2.2-20231207143243-e49617d2e4fe.tm.json",
+					Digest:    "e49617d2e4fe",
+					TimeStamp: "20231207143243",
+				},
+				FoundIn: model.FoundSource{RemoteName: "r1"},
+			},
+			{
+				TOCVersion: model.TOCVersion{
+					Version:   model.Version{Model: "v1.2.3"},
+					TMID:      "author/manufacturer/mpn/v1.2.3-20231207153243-e49617d2e4ff.tm.json",
+					Digest:    "e49617d2e4ff",
+					TimeStamp: "20231207153243",
+				},
+				FoundIn: model.FoundSource{RemoteName: "r1"},
+			},
+			{
+				TOCVersion: model.TOCVersion{
+					Version:   model.Version{Model: "v2.0.0"},
+					TMID:      "author/manufacturer/mpn/v2.0.0-20231208123243-f49617d2e4fc.tm.json",
+					Digest:    "f49617d2e4fc",
 					TimeStamp: "20231205123243",
 				},
 				FoundIn: model.FoundSource{RemoteName: "r1"},